# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Front API",
        "version": "0.1"
    },
    "paths": {
...
        "api.motovago.com/categories": {
            "post": {
                "summary": "Gets all categories in database",
                "operationId": "categories",
                "parameters": [
                    {
                        "name": "location_guid",
                        "in": "query",
                        "description": "location_guid",
                        "required": true,
                        "example": "123123-4123141-23123123"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "200"
                                        },
                                        "categories": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "12"
                                                    },
                                                    "category_guid": {
                                                        "type": "string",
                                                        "example": "d4adbd65-9903-4df0-aaaf-2aa491e99123"
                                                    },
                                                    "name_en": {
                                                        "type": "string",
                                                        "example": "Cars"
                                                    },
                                                    "name_ar": {
                                                        "type": "string",
                                                        "example": "سيارات"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "cars"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "https://api.motovago.test/storage/images/locations/pc_photo"
                                                    },
                                                    "queue": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "rent_listings": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "string",
                                                                    "example": "5"
                                                                },
                                                                "listing_guid": {
                                                                    "type": "string",
                                                                    "example": "a43013a8-dd46-4055-b9df-9024c1092169"
                                                                },
                                                                "category_guid": {
                                                                    "type": "string",
                                                                    "example": "a43013a8-4055-b9df-9024c1092169"
                                                                },
                                                                "subcategory_guid": {
                                                                    "type": "string",
                                                                    "example": "232323-4055-b9df-9024c1092169"
                                                                },
                                                                "listing_no": {
                                                                    "type": "string",
                                                                    "example": "47382643"
                                                                },
                                                                "price": {
                                                                    "type": "string",
                                                                    "example": "500.000"
                                                                },
                                                                "main_image_api": {
                                                                    "type": "array",
                                                                    "items": {
                                                                        "properties": {
                                                                            "listing_guid": {
                                                                                "type": "string",
                                                                                "example": "6f1ceeac-9719-43ec-9403-b4f8bbbc0adbc"
                                                                            },
                                                                            "name": {
                                                                                "type": "string",
                                                                                "example": "https://motovago.com/storage/listings/78603483/2.jpeg"
                                                                            }
                                                                        },
                                                                        "type": "object",
                                                                        "format": "query"
                                                                    }
                                                                }
                                                            },
                                                            "type": "object",
                                                            "format": "query"
                                                        }
                                                    },
                                                    "sale_listings": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "string",
                                                                    "example": "6"
                                                                },
                                                                "listing_guid": {
                                                                    "type": "string",
                                                                    "example": "232332-dd46-4055-b9df-9024c1092169"
                                                                },
                                                                "category_guid": {
                                                                    "type": "string",
                                                                    "example": "121212-4055-b9df-9024c1092169"
                                                                },
                                                                "subcategory_guid": {
                                                                    "type": "string",
                                                                    "example": "4141414-4055-b9df-9024c1092169"
                                                                },
                                                                "name_en": {
                                                                    "type": "string",
                                                                    "example": "BORUSAN ÇIKIŞLI 2018 MODEL RANGE ROVER SPORT HATASIZ 69000 KM"
                                                                },
                                                                "name_ar": {
                                                                    "type": "string",
                                                                    "example": "null"
                                                                },
                                                                "listing_no": {
                                                                    "type": "string",
                                                                    "example": "11111"
                                                                },
                                                                "price": {
                                                                    "type": "string",
                                                                    "example": "233.000"
                                                                },
                                                                "main_image_api": {
                                                                    "type": "array",
                                                                    "items": {
                                                                        "properties": {
                                                                            "listing_guid": {
                                                                                "type": "string",
                                                                                "example": "6f1ceeac-9719-43ec-9403-b4f8bbbc0adc"
                                                                            },
                                                                            "name": {
                                                                                "type": "string",
                                                                                "example": "https://motovago.com/storage/listings/78603483/1.jpeg"
                                                                            }
                                                                        },
                                                                        "type": "object",
                                                                        "format": "query"
                                                                    }
                                                                }
                                                            },
                                                            "type": "object",
                                                            "format": "query"
                                                        }
                                                    },
                                                    "listing_count": {
                                                        "type": "string",
                                                        "example": "8"
                                                    }
                                                },
                                                "type": "object",
                                                "format": "query"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "api.motovago.com/getLocations": {
            "get": {
                "summary": "Gets all locations in locations table where status is '1'",
                "operationId": "locations",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "image": {
                                            "type": "string",
                                            "example": "https://api.motovago.test/storage/images/locations/pc_photo/qatar.jpeg"
                                        },
                                        "location_guid": {
                                            "type": "string",
                                            "example": "03d6cbbb-a43f-4143-8f9c-288485de12e4"
                                        },
                                        "name_en": {
                                            "type": "string",
                                            "example": "QATAR"
                                        },
                                        "name_ar": {
                                            "type": "string",
                                            "example": "دولة قطر"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "400"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Error occured while getting locations list."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}